adjust for specific wisco cases; move code to new file
